% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptional.R
\name{create_exception}
\alias{create_exception}
\title{Create an exception}
\usage{
create_exception(message, nature, ...)
}
\description{
Create an objecto fo class "exceptional" that contains a message and data
that can be further manipulated programatically.
}
