% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptional.R
\name{create_message}
\alias{create_message}
\title{Create user messages}
\usage{
create_message(default, ...)
}
\arguments{
\item{default}{The default message as a string if no error message customized
for the language local can be found.}

\item{...}{further named arugments. The name must correspond to the
[ISO-639-1](https://en.wikipedia.org/wiki/ISO_639-1) abbreviation (like
hu for Hungary), the value is the message as a string.}
}
\description{
Create user messages
}
