% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature.R
\name{create_nature}
\alias{create_nature}
\title{Create one of the allowed natures}
\usage{
create_nature(nature, allowed = lookup_allowed_natrues())
}
\arguments{
\item{nature}{A character string of length one indicating the desired
nature of the exception}

\item{allowed}{an ordered, empty factor that contains the levels that are
allowed.}
}
\description{
Create one of the allowed natures
}
