% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptional.R
\name{create_exception}
\alias{create_exception}
\title{Create an exception}
\usage{
create_exception(message, nature, id = NULL, ..., allowed_natures = NULL)
}
\arguments{
\item{message}{The message that should be displayed to the user. Object of
class "exceptional-message", created with [create_message()].}

\item{nature}{The nature of the exception.
`allowed_natures`.}
}
\description{
Create an objecto fo class "exceptional" that contains a message and data
that can be further manipulated programatically.
}
